Adapt the whole config file to 1.20.1 config value types and fix the .back common config file creation Refactor the main file for clarity and logic in calling mod parts Fix the server error "p_130086 is null" by correcting recipe serializer register

+++ src/main/java/sfiomn/legendarysurvivaloverhaul/api/bodydamage/MalusBodyPartEnum.java
@@ -20,9 +20,9 @@ public enum MalusBodyPartEnum {
 
     public final List<? extends String> effects;
     public final List<? extends Integer> amplifiers;
-    public final List<? extends Float> thresholds;
+    public final List<? extends Double> thresholds;
 
-    MalusBodyPartEnum(List<? extends String> effects, List<? extends Integer> amplifiers, List<? extends Float> thresholds) {
+    MalusBodyPartEnum(List<? extends String> effects, List<? extends Integer> amplifiers, List<? extends Double> thresholds) {
         this.effects = effects;
         this.amplifiers = amplifiers;
         this.thresholds = thresholds;
